requirements.txt
PyPDF2
Pillow
windows-curses; sys_platform == 'win32'
InquirerPy